Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 129ms, Memory: 10MB, Status: Passed - Codetree

+++ 250722/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -24,13 +24,13 @@ public class Main {
 	
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        long m = sc.nextInt();
+        long m = sc.nextLong();
         
         int minCnt = Integer.MAX_VALUE;
         int maxCnt = Integer.MIN_VALUE;
         
-        long a = sc.nextInt();
-        long b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
         
         for(long target = a; target<=b; target++) {
         	int tempCnt = binarySearch(m, target);
